Ready codebase for merge into Arrow C++ codebase

+++ src/parquet/arrow/record_reader.cc
@@ -532,7 +532,7 @@ inline void TypedRecordReader<ByteArrayType>::ReadValuesDense(int64_t values_to_
   auto builder = static_cast<::arrow::BinaryBuilder*>(builder_.get());
   for (int64_t i = 0; i < num_decoded; i++) {
     PARQUET_THROW_NOT_OK(
-        builder->Append(values[i].ptr, static_cast<int64_t>(values[i].len)));
+        builder->Append(values[i].ptr, static_cast<int32_t>(values[i].len)));
   }
   ResetValues();
 }
@@ -568,7 +568,7 @@ inline void TypedRecordReader<ByteArrayType>::ReadValuesSpaced(int64_t values_to
   for (int64_t i = 0; i < num_decoded; i++) {
     if (::arrow::BitUtil::GetBit(valid_bits, valid_bits_offset + i)) {
       PARQUET_THROW_NOT_OK(
-          builder->Append(values[i].ptr, static_cast<int64_t>(values[i].len)));
+          builder->Append(values[i].ptr, static_cast<int32_t>(values[i].len)));
     } else {
       PARQUET_THROW_NOT_OK(builder->AppendNull());
     }
